pyproject.toml
[project]
name = "job-hunter-mcp"
version = "0.2.0"
description = "AI Job Hunter MCP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.12.0",
    "openai>=1.97.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "fastapi>=0.116.1",
    "sqlalchemy>=2.0.41",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "toml>=0.10.2",
]
